requirements.txt
setuptools~=80.3.1
numpy~=2.0.2
pandas~=2.3.1
requests~=2.32.4
finnhub-python~=2.4.24
